Add files via upload
This implements Angelo's updated UI, as well as the updated authentication.
To register for a new profile, select 'create one' and then it will prompt you to add a new profile to the database. Something to continue work on will be to extend authentication to verify users are within the .syr domain.